Dodana djelomicna backend funkcionalnost za dodavanje prijave, potrebna dorada. Dodana manytomany relacija koja se izgubila u međuvremenu, dodana dinamika na frontendu

+++ AutoServisProject/AutoServis/src/main/java/opp/dao/RadnoVrijemeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository("radnoVrijemeRepository")
-public interface RadnoVrijemeRepository extends JpaRepository<RadnoVrijeme, Long> {
+public interface RadnoVrijemeRepository extends JpaRepository<RadnoVrijeme, Integer> {
 
 }
